services/ingestion_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
python-dotenv
pydantic
pytest
httpx